refactor(formats): split gene-model domain modules
